+++ R/globals.R
@@ -1,1 +1,4 @@
-utils::globalVariables(c("label", "value", "x", "xend", "y", "yend", "zone"))
+utils::globalVariables(c(
+  "column", "display", "label", "value", "x", "xend", "xmax", "xmin",
+  "y", "yend", "ymax", "ymin", "zone"
+))